[issues #22] Fixed a zone value type.

+++ grs/cttwt.py
@@ -29,7 +29,7 @@ class TWTime(object):
   '''
   def __init__(self,tz = 8):
     try:
-      self.TimeZone = int(tz)
+      self.TimeZone = float(tz)
     except:
       self.TimeZone = 8
